refactor(FinalRoundTitleChanger): gave component its own file
